test(persistence): move effector config tests to sibling file
